requirements.txt
django-ckeditor-updated==4.4.4
South==1.0
